+++ src/styles.scss
@@ -3,7 +3,7 @@
 @tailwind components;
 @tailwind utilities;
 
-@import '~@angular/cdk/overlay-prebuilt.css';
+@import "~@angular/cdk/overlay-prebuilt.css";
 
 @import url("https://fonts.googleapis.com/css2?family=Marcellus+SC&family=Ovo&display=swap");
 @import url("https://fonts.googleapis.com/css2?family=Marcellus+SC&family=Ovo&display=swap");
